style: rename `api` to `cli`
