+++ src/imap_mag/cli/cliUtils.py
