+++ src/imap_mag/cli/fetch/binary.py
@@ -6,7 +6,7 @@ from typing import Annotated, Optional
 import typer
 
 from imap_mag import appUtils
-from imap_mag.api.apiUtils import initialiseLoggingForCommand
+from imap_mag.cli.cliUtils import initialiseLoggingForCommand
 from imap_mag.client.WebPODA import WebPODA
 from imap_mag.config import AppSettings, FetchMode
 from imap_mag.download.FetchBinary import FetchBinary
