+++ src/imap_mag/cli/process.py
@@ -5,7 +5,7 @@ from typing import Annotated
 import typer
 
 from imap_mag import appUtils
-from imap_mag.api.apiUtils import (
+from imap_mag.cli.cliUtils import (
     fetch_file_for_work,
     initialiseLoggingForCommand,
 )
